feat: add amdgpu firmware
AMDgpu is an open source device driver for the Linux operating system developed by AMD
to support its Radeon lineup of graphics cards (GPUs). It was announced in 2014 as
the successor to the previous radeon device driver as part of AMD's new "unified"
driver strategy, and was released on April 20, 2015.

Fixes: https://github.com/siderolabs/pkgs/issues/792

Signed-off-by: Victor Seva <linuxmaniac@torreviejawireless.org>
Signed-off-by: Noel Georgi <git@frezbo.dev>