feat: update stargz extension
This PR updates the stargz extension to pull in busybox, exec into the pid 1 mount namespace, then start the stargz snapshotter. This allows us to work around what seems to be mount propagation issues with extensions. Once those are better understood and fixed within Talos, this extenstion can hopefully be updated to remove the namespace switching.

Signed-off-by: Spencer Smith <spencer.smith@talos-systems.com>