feat: add zfs-service to zfs extension (unmount, encryption)
This patch adds a new service to the zfs extension (`zfs-service`) that
handles pool import and unmount. These operations are tied to the
service lifecycle: the service imports all pools when it starts, waits
for an exit signal, then unmounts all pools before exiting.

As a subtle additional benefit, the service passes the `-l` flag to
`zpool-import`[^1], which instructs it to request encryption keys for
all encrypted datasets. Using the zfs `keylocation`[^2] property and
Talos secure boot and TPM disk encryption, a ZFS encryption key file can
be safely stored on the EPHEMERAL partition to import encrypted datasets
at boot. Alternatively, a key can be stored on an https server.

[^1]: https://openzfs.github.io/openzfs-docs/man/master/8/zpool-import.8.html
[^2]: https://openzfs.github.io/openzfs-docs/man/master/7/zfsprops.7.html#keylocation

Signed-off-by: Jean-Francois Roy <jf@devklog.net>
Signed-off-by: Noel Georgi <git@frezbo.dev>