fix: unable to override runtime defaults
This fixes the issue when `.machine.files` using
`/etc/cri/conf.d/20-customization.part` is sometimes unable to merge the
values.

The fix is to prefix the the files with `10-` prefix so the
customization with `20-` takes precedence when merging.

`gvisor-debug` is prefixed with `11-` since it should overwrite the
gvisor defaults. This also fixes where the order of gvisor extensions
mattered previously.

Fixes: #578

Signed-off-by: Noel Georgi <git@frezbo.dev>
(cherry picked from commit 9c92bda95b12c7c00a585778423c8ec8129465c5)