fix(glibc): use lib64 for better compatibility
This patch changes the glibc extension to use the `lib64` directory
for libraries instead of `lib`. This better reflects a standard
Fedora/FHS/LSB distribution and fixes issues with a few containers, with
the Nvidia Linux installer, and with the Nvidia container toolkit.

The Nvidia container toolkit mounts libraries at the same path as the
host system and runs `ldconfig` to update the cache. Mounting 64-bit
libraries in `lib` is wrong on Fedora-like _and_ Ubuntu-like
distributions.

Signed-off-by: Jean-Francois Roy <jf@devklog.net>
Signed-off-by: Noel Georgi <git@frezbo.dev>