feat(thunderbolt): disable auto power management
I bet most people who enable this extension are using Thunderbolt in a
manner that requires it to always be on and not have type of "auto"
power mode where it could be turned on or off to save power.

For me, I am running into issues with using thunderbolt for a Ceph ring
and I've noticed every so often that the thunderbolt interfaces are
dropped on node loss which requires me to replug the thunderbolt cable
or reboot my cluster nodes.

It would be great if the thunderbolt extension could include this by
default, since infrastructure/server use cases generally need stable,
always-on connectivity rather than power-saving behavior.

Signed-off-by: Devin Buhl <onedr0p@users.noreply.github.com>
Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>